课件及杂项/课件及老师发的文件/bibliometrics-mini-template/bibliometrics-mini-template/requirements.txt
pandas>=2.2
numpy>=1.26
scipy>=1.11
networkx>=3.2
matplotlib>=3.8
PyYAML>=6.0
requests>=2.31
tqdm>=4.66
python-dotenv>=1.0
pytest>=8.0
